+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - The slugrunner doesn't mount the storage keys anymore. An init container is
   responsible for downloading the slug
 - Bump the default slugrunner version to v3.0.1
+- Change Method `CreateSecret` to `CreateOrUpdateSecret` on k8s interfaces
 
 ## [0.15.0] - 2018-02-14
 ### Changed